+++ dinghy-lib/src/apple/platform.rs
@@ -45,22 +45,14 @@ impl AppleDevicePlatform {
 
     fn sysroot_path(&self) -> Result<String> {
         let sdk_name = match self.sim {
-            Some(AppleSimulatorType::Ios) => {
-                "iphonesimulator"
-            }
-            Some(AppleSimulatorType::Tvos) => {
-                "appletvsimulator"
-            }
-            Some(AppleSimulatorType::Watchos) => {
-                "watchsimulator"
-            }
-            Some(AppleSimulatorType::Visionos) => {
-                // xros and xrsimulator are sdk names.
-                // See https://github.com/rust-lang/rust/pull/121419#discussion_r1501908152 for
-                // more inconsistencies about visionOS and xrOS.
-                "xrsimulator"
-            }
-            None => "iphoneos"
+            Some(AppleSimulatorType::Ios) => "iphonesimulator",
+            Some(AppleSimulatorType::Tvos) => "appletvsimulator",
+            Some(AppleSimulatorType::Watchos) => "watchsimulator",
+             // xros and xrsimulator are sdk names.
+             // See https://github.com/rust-lang/rust/pull/121419#discussion_r1501908152 for
+             // more inconsistencies about visionOS and xrOS.
+             Some(AppleSimulatorType::Visionos) => "xrsimulator",
+            None => "iphoneos",
         };
         let xcrun = process::Command::new("xcrun")
             .args(&["--sdk", sdk_name, "--show-sdk-path"])